zookeeper_amqp/ms/requirements.txt
ipython
mysql-connector
pika
kazoo
docker